Adjust hash functions and HashBits
